feat: move isValidSelection into util; write first test for stringify
