+++ packages/discovery-react-components/src/components/StructuredQuery/utils/index.ts
@@ -1,8 +1,8 @@
 export { getFieldNames } from './getFieldNames';
 export { getNewId } from './getNewId';
+export { isValidSelection } from './isValidSelection';
 export { stringifyStructuredQuerySelection } from './stringifyStructuredQuerySelection';
 export {
-  Row,
   Group,
   StructuredQuerySelection,
   OperatorDropdownSelectedItem,
